+++ fuzzers/selectfuzz/build_drivers/poppler/pdfimages-instrument.sh
@@ -15,6 +15,15 @@ export TMP_DIR=$OUT/tmp-$1
 export CFLAGS="$CFLAGS -distance=$TMP_DIR/distance.cfg.txt"
 export CXXFLAGS="$CXXFLAGS -distance=$TMP_DIR/distance.cfg.txt"
 
+
+if [ ! -d "$TARGET/repo" ]; then
+    echo "fetch.sh must be executed first."
+    exit 1
+fi
+
+git config --global --add safe.directory /magma/targets/poppler/freetype2
+git config --global --add safe.directory /magma/targets/poppler/repo
+
 export WORK="$TARGET/work"
 rm -rf "$WORK"
 mkdir -p "$WORK"
@@ -35,6 +44,7 @@ EXTRA=""
 test -n "$AR" && EXTRA="$EXTRA -DCMAKE_AR=$AR"
 test -n "$RANLIB" && EXTRA="$EXTRA -DCMAKE_RANLIB=$RANLIB"
 
+export LDFLAGS="$LDFLAGS -lbrotlidec"
 cmake "$TARGET/repo" \
   $EXTRA \
   -DCMAKE_BUILD_TYPE=debug \
@@ -67,6 +77,6 @@ cp "$WORK/poppler/utils/"{pdfimages,pdftoppm} "$OUT/"
 $CXX $CXXFLAGS -std=c++11 -I"$WORK/poppler/cpp" -I"$TARGET/repo/cpp" \
     "$TARGET/src/pdf_fuzzer.cc" -o "$OUT/pdf_fuzzer" \
     "$WORK/poppler/cpp/libpoppler-cpp.a" "$WORK/poppler/libpoppler.a" \
-    "$WORK/lib/libfreetype.a" $LDFLAGS $LIBS -ljpeg -lz \
+    "$WORK/lib/libfreetype.a" $LDFLAGS $LIBS $FUZZER_LIB -ljpeg -lz \
     -lopenjp2 -lpng -ltiff -llcms2 -lm -lpthread -pthread
